myapp/requirements.txt
Flask==2.5.2
Flask-Login==0.6.2
Flask-SQLAlchemy==2.5.1
Flask-Migrate==3.1.0
Flask-WTF==1.0.0
Flask-Mail==0.9.1
WTForms==3.0.1
PyYAML==6.0
Werkzeug==2.5.2
SQLAlchemy==1.4.47
itsdangerous==2.2.0
reportlab==3.6.2
